refactor: features/profile index.ts 공개면 정리 및 app/ deep import 교체
