+++ app/(protected)/profile/[userId]/project/[projectId]/page.tsx
@@ -1,7 +1,6 @@
 import { dehydrate } from '@tanstack/react-query';
 
-import { fetchMyProfile } from '@/features/profile/api/fetch-my-profile';
-import { profileKeys } from '@/features/profile/profile.query-keys';
+import { fetchMyProfile, profileKeys } from '@/features/profile';
 import { fetchProject } from '@/features/project/api/fetchers';
 import { projectKeys } from '@/features/project/api/query-keys';
 import { ProjectDetailStream } from '@/features/project/components/project-detail-stream';
